refactor infoHead component
- move InfoHead to `components`folder
- move logic from InfoHeadFilled to InfoHead component
